Restructure of wrapper module
