+++ extendr-api/src/prelude.rs
@@ -30,7 +30,7 @@ pub use crate::{
 
 pub use super::logical::Bool;
 
-pub use super::matrix::{RArray, RColumn, RMatrix, RMatrix3D};
+pub use super::wrapper::{RArray, RColumn, RMatrix, RMatrix3D};
 
 pub use super::robj::{IntoRobj, Robj, RobjItertools};
